Remove Parse library and supporting code
