+++ VideoLocker/build.gradle
@@ -98,10 +98,6 @@ dependencies {
     //New Relic Integration
     compile 'com.newrelic.agent.android:android-agent:5.4.1'
 
-    // Parse notification service
-    compile 'com.parse:parse-android:1.13.0'
-    compile 'com.parse.bolts:bolts-android:1.4.0'
-
     //Include DI framework
     compile 'org.roboguice:roboguice:3.0.1'
     provided 'org.roboguice:roboblender:3.0.1'
