+++ VideoLocker/src/main/java/org/edx/mobile/base/MainApplication.java
@@ -13,8 +13,6 @@ import com.google.inject.Module;
 import com.joanzapata.iconify.Iconify;
 import com.joanzapata.iconify.fonts.FontAwesomeModule;
 import com.newrelic.agent.android.NewRelic;
-import com.parse.Parse;
-import com.parse.ParseInstallation;
 
 import org.edx.mobile.BuildConfig;
 import org.edx.mobile.R;
@@ -98,22 +96,7 @@ public abstract class MainApplication extends MultiDexApplication {
             com.facebook.Settings.setApplicationId(config.getFacebookConfig().getFacebookAppId());
         }
 
-        boolean needVersionUpgrade = needVersionUpgrade(this);
-        // initialize Parse notification
-        // it maybe good to support multiple notification providers running
-        // at the same time, as it is less like to be the case in the future,
-        // we at two level of controls just for easy change of different providers.
-        if (config.isNotificationEnabled()) {
-            Config.ParseNotificationConfig parseNotificationConfig =
-                    config.getParseNotificationConfig();
-            if (parseNotificationConfig.isEnabled()) {
-                Parse.enableLocalDatastore(this);
-                Parse.initialize(this, parseNotificationConfig.getParseApplicationId(), parseNotificationConfig.getParseClientKey());
-                tryToUpdateParseForAppUpgrade(this, needVersionUpgrade);
-            }
-        }
-
-        if (needVersionUpgrade) {
+        if (needVersionUpgrade(this)) {
             // try repair of download data if app version is updated
             injector.getInstance(IStorage.class).repairDownloadCompletionData();
         }
@@ -149,47 +132,6 @@ public abstract class MainApplication extends MultiDexApplication {
             pmanager.setAppVersionCode(curVersion);
         }
         return needVersionUpgrade;
-    }
-
-    /**
-     * if app is launched from upgrading, we need to resync with parse server.
-     *
-     * @param context
-     */
-    private void tryToUpdateParseForAppUpgrade(Context context, boolean needVersionUpgrade) {
-
-        PrefManager.AppInfoPrefManager pmanager = new PrefManager.AppInfoPrefManager(context);
-        boolean hadNotification = pmanager.isNotificationEnabled();
-        if (needVersionUpgrade) {
-            if (hadNotification) {
-                pmanager.setAppUpgradeNeedSyncWithParse(true);
-            }
-        }
-        ParseInstallation installation = ParseInstallation.getCurrentInstallation();
-        final String languageKey = "preferredLanguage";
-        final String countryKey = "preferredCountry";
-        String savedPreferredLanguage = installation.getString(languageKey);
-        String savedPreferredCountry = installation.getString(countryKey);
-        Locale locale = Locale.getDefault();
-        String currentPreferredLanguage = locale.getLanguage();
-        String currentPreferredCountry = locale.getCountry();
-        boolean dirty = false;
-        if (!currentPreferredLanguage.equals(savedPreferredLanguage)) {
-            installation.put(languageKey, currentPreferredLanguage);
-            dirty = true;
-        }
-        if (!currentPreferredCountry.equals(savedPreferredCountry)) {
-            installation.put(countryKey, currentPreferredCountry);
-            dirty = true;
-        }
-        if (dirty) {
-            try {
-                installation.saveInBackground();
-            } catch (Exception ex) {
-                logger.error(ex);
-            }
-        }
-        pmanager.setNotificationEnabled(true);
     }
 
     public static class CrashlyticsCrashReportObserver {
